requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.4.2
PIL==1.1.7
Whoosh==2.4.1
argparse==1.2.1
django-appconf==0.5
django-forms-bootstrap==2.0.3.post1
django-user-accounts==1.0b4
metron==1.0
pinax-theme-bootstrap==2.0.4
pinax-theme-bootstrap-account==1.0b2
pinax-utils==1.0b1.dev3
pytz==2012d
wsgiref==0.1.2
dj-database-url==0.2.0
psycopg2
